🚸 Config wizard: Use an homebrewed function for questions
`inquirer` does not play well with long questions. I made sure to use the same "style" of questions as to keep as much consistency as possible while taking advantage of the fact that I made the function to provide better prompt markers (--argument-name=).
Those prompt markers are on their own lines, where as inputing potentially long text with inquirer's prompts is troublesome (the input is on the same line as the question)