💄 Improve styling of colored cards (gkeep)
Instead of giving the card a black background and coloring the text some
color, remove the border (because border on background is ugly) and make
the background the card's color. The foreground color is determined with
`readable_on`. Also update `readable_on` to return Rich colors by
default, since it's what I'm using everywhere.